backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
streamlit
requests
gradio
google-generativeai